% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self_suff_standard.R
\name{calculate_food_cost}
\alias{calculate_food_cost}
\title{Calculates food costs for each family based on age groups and food plans}
\usage{
calculate_food_cost(food_df, sss_df)
}
\arguments{
\item{food_df}{A dataframe containing food cost data for different age groups and plans.}

\item{sss_df}{A dataframe containing self-sufficiency data for families.}
}
\value{
A dataframe with an additional column for the calculated food plan cost.
}
\description{
This function calculates the total food cost for each family in the self-sufficiency
dataframe based on the number of people in each age group and the selected food plan.
}
