% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_data.R
\name{download_job_salary_data}
\alias{download_job_salary_data}
\title{Downloads job salary data from the BLS website}
\usage{
download_job_salary_data()
}
\value{
None. Downloads the file and moves it to the \code{DataFiles/RawOutputFiles} directory.
}
\description{
This function uses Selenium to navigate to the BLS website and download the
job salary data file. The file is then moved to the specified directory.
}
