% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self_suff_standard.R
\name{food_cost_test}
\alias{food_cost_test}
\title{Calculates food costs for all families and food plans}
\usage{
food_cost_test(sss_df, food_plans_df)
}
\arguments{
\item{sss_df}{A dataframe containing self-sufficiency data for families.}

\item{food_plans_df}{A dataframe containing food cost data for different age groups and plans.}
}
\value{
A dataframe with an additional column for the calculated food plan cost.
}
\description{
This function calculates the total food cost for all families in the self-sufficiency
dataframe across all food plans by matching age groups and food plan costs.
}
