% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food_plans.R
\name{food_plans_main}
\alias{food_plans_main}
\title{Fetches and processes the Food Plans data}
\usage{
food_plans_main()
}
\value{
A fused dataframe containing processed data from the Thrifty and Low to Liberal Food Plans.
#' @export
}
\description{
This function scrapes the USDA Food Plans website, downloads the relevant PDF files,
and processes the data to extract monthly cost information for different food plans.
}
