% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_search.R
\name{get_5_jobs}
\alias{get_5_jobs}
\title{Retrieves the top 5 jobs for a specific county based on target salary}
\usage{
get_5_jobs(county)
}
\arguments{
\item{county}{A string representing the name of the county.}
}
\value{
None. Prints the top 5 jobs with salaries near or above the target.
}
\description{
This function retrieves the target yearly salary based on family type, housing,
and food plan, and then finds the top 5 jobs in the specified county with salaries
near or above the target.
}
