% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_data.R
\name{get_education_requirement}
\alias{get_education_requirement}
\title{Retrieves education requirements for the given job data}
\usage{
get_education_requirement(df)
}
\arguments{
\item{df}{A dataframe containing job data.}
}
\value{
A dataframe with education requirements added.
}
\description{
This function reads a file containing education requirements and merges it
with the provided job data to add education requirements.
}
