% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self_suff_standard.R
\name{get_final_self_suff_standard}
\alias{get_final_self_suff_standard}
\title{Generates the final Self Sufficiency Standard data}
\usage{
get_final_self_suff_standard()
}
\value{
None. Writes the final self-sufficiency data to \code{self_suff_standard.xlsx}.
}
\description{
This function combines raw self-sufficiency data with housing and food costs,
calculates monthly and yearly costs, and writes the final data to an Excel file.
}
