% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food_plans.R
\name{get_food_plans}
\alias{get_food_plans}
\title{Generates summarized Food Plans data}
\usage{
get_food_plans()
}
\value{
None. Writes the summarized data to an Excel file.
}
\description{
This function calculates the mean monthly costs for each food plan by age group
and writes the summarized data to an Excel file.
}
