% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food_plans.R
\name{get_fused_dfs}
\alias{get_fused_dfs}
\title{Combines data from the Thrifty and Low to Liberal Food Plans}
\usage{
get_fused_dfs(df1, df2)
}
\arguments{
\item{df1}{A tibble containing Thrifty Food Plan data.}

\item{df2}{A tibble containing Low to Liberal Food Plan data.}
}
\value{
A tibble containing the combined data.
}
\description{
This function fuses the data from the Thrifty Food Plan and the Low to Liberal Food Plan
into a single dataframe with additional metadata.
}
