% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housing_cost.R
\name{get_housing_cost}
\alias{get_housing_cost}
\title{Writes housing cost data to an Excel file}
\usage{
get_housing_cost(fips_code)
}
\arguments{
\item{fips_code}{A string representing the FIPS code of the county.}
}
\value{
None. Writes the housing cost data to an Excel file named \code{housing_cost.xlsx}.
}
\description{
This function retrieves housing cost data for a specified county using its FIPS code
and writes the data to an Excel file.
}
