% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_data.R
\name{get_job_data}
\alias{get_job_data}
\title{Processes job salary data for a specific county and state}
\usage{
get_job_data(county_id, state_id, state_name, state_abreviation, county_name)
}
\arguments{
\item{county_id}{A string representing the county ID.}

\item{state_id}{A string representing the state ID.}

\item{state_name}{A string representing the name of the state.}

\item{state_abreviation}{A string representing the state abbreviation.}

\item{county_name}{A string representing the name of the county.}
}
\value{
A tibble containing processed job salary data.
}
\description{
This function reads the job salary data file, filters it for the specified
county and state, and processes it to calculate mean wages.
}
