% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_data.R
\name{get_job_data_and_ed_req}
\alias{get_job_data_and_ed_req}
\title{Retrieves job data and education requirements for a specific county and state}
\usage{
get_job_data_and_ed_req(county_name, state_abreviation, state_name)
}
\arguments{
\item{county_name}{A string representing the name of the county.}

\item{state_abreviation}{A string representing the state abbreviation.}

\item{state_name}{A string representing the name of the state.}
}
\value{
None. Writes the data to an Excel file.
}
\description{
This function retrieves job data and education requirements for the specified
county and state, and writes the data to an Excel file.
}
