% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_data.R
\name{get_job_titles}
\alias{get_job_titles}
\title{Retrieves job titles for the given job data}
\usage{
get_job_titles(df)
}
\arguments{
\item{df}{A dataframe containing job data.}
}
\value{
A dataframe with job titles added.
}
\description{
This function reads a file containing job titles and merges it with the
provided job data to add job titles.
}
