% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_search.R
\name{get_jobs}
\alias{get_jobs}
\title{Retrieves jobs with salaries greater than or near the target salary}
\usage{
get_jobs(target, county)
}
\arguments{
\item{target}{A numeric value representing the target annual salary.}

\item{county}{A string representing the name of the county.}
}
\value{
None. Prints the top 5 jobs with salaries near or above the target.
}
\description{
This function filters job data for a specific county to find jobs with annual
salaries greater than or near the specified target salary. It returns the top 5
jobs closest to the target salary.
}
