% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food_plans.R
\name{get_low_to_lib}
\alias{get_low_to_lib}
\title{Processes the Low to Liberal Food Plan PDF}
\usage{
get_low_to_lib()
}
\value{
A tibble containing the monthly cost data for the Low, Moderate, and Liberal Food Plans.
}
\description{
This function extracts and processes data from the Low to Liberal Food Plan PDF.
}
