% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self_suff_standard.R
\name{get_monthly_cost}
\alias{get_monthly_cost}
\title{Calculates monthly and yearly costs for each family}
\usage{
get_monthly_cost(sss_df)
}
\arguments{
\item{sss_df}{A dataframe containing self-sufficiency data with cost components.}
}
\value{
A dataframe with additional columns for monthly and yearly costs.
}
\description{
This function calculates the total monthly and yearly costs for each family
by summing up various cost components, including housing and food costs.
}
