% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_data.R
\name{get_place_id}
\alias{get_place_id}
\title{Retrieves the place ID for a given county}
\usage{
get_place_id(county)
}
\arguments{
\item{county}{A string representing the name of the county.}
}
\value{
A string representing the place ID for the county.
}
\description{
This function reads a file containing county codes and extracts the place ID
for the specified county.
}
