% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_self_suff_standard.R
\name{get_raw_self_suff_standard}
\alias{get_raw_self_suff_standard}
\title{Retrieves and writes the Self Sufficiency Standard data to an Excel file}
\usage{
get_raw_self_suff_standard(county_name, state_name)
}
\arguments{
\item{county_name}{A string representing the name of the county.}

\item{state_name}{A string representing the name of the state.}
}
\value{
None. Writes the processed data to \code{Raw_self_suff_standard.xlsx}.
}
\description{
This function fetches the Self Sufficiency Standard data for the specified county and state,
processes it, and writes the resulting data to an Excel file.
}
