% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self_suff_standard.R
\name{get_sss}
\alias{get_sss}
\title{Combines raw self-sufficiency data with housing and food costs}
\usage{
get_sss(raw_sss, housing_df, food_df)
}
\arguments{
\item{raw_sss}{A dataframe containing raw self-sufficiency data.}

\item{housing_df}{A dataframe containing housing cost data.}

\item{food_df}{A dataframe containing food cost data.}
}
\value{
A dataframe containing the combined self-sufficiency data with calculated costs.
}
\description{
This function merges the raw self-sufficiency data with housing and food cost data,
calculates the monthly and yearly costs, and returns the final dataframe.
}
