% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_data.R
\name{get_state_id}
\alias{get_state_id}
\title{Retrieves the state ID for a given state}
\usage{
get_state_id(state)
}
\arguments{
\item{state}{A string representing the name of the state.}
}
\value{
A string representing the state ID for the state.
}
\description{
This function reads a file containing state codes and extracts the state ID
for the specified state.
}
