% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_search.R
\name{get_target_by_index}
\alias{get_target_by_index}
\title{Retrieves the target yearly salary for a specific index}
\usage{
get_target_by_index(index)
}
\arguments{
\item{index}{An integer representing the index of the desired row in the data.}
}
\value{
A numeric value representing the target yearly salary.
}
\description{
This function retrieves the yearly cost (target salary) for a specific index
from the self-sufficiency standard data.
}
