% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_search.R
\name{get_target_table_cols}
\alias{get_target_table_cols}
\title{Retrieves the target yearly salary based on family type, housing, and food plan}
\usage{
get_target_table_cols()
}
\value{
A numeric value representing the target yearly salary.
}
\description{
This function filters the self-sufficiency standard data based on the specified
family type, housing type, and food plan to retrieve the target yearly salary.
}
