% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food_plans.R
\name{get_thrifty}
\alias{get_thrifty}
\title{Processes the Thrifty Food Plan PDF}
\usage{
get_thrifty()
}
\value{
A tibble containing the monthly cost data for the Thrifty Food Plan.
}
\description{
This function extracts and processes data from the Thrifty Food Plan PDF.
}
