% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housing_cost.R
\name{housing_cost_main}
\alias{housing_cost_main}
\title{Retrieves housing cost data for a specified county}
\usage{
housing_cost_main(county_code)
}
\arguments{
\item{county_code}{A string representing the FIPS code of the county.}
}
\value{
A tibble containing housing cost data for different housing types (e.g., Efficiency, One-Bedroom).
Returns \code{NULL} if the API request fails.
}
\description{
This function makes an API request to the HUD website to retrieve the most recent
housing cost data for a specified county using its FIPS code.
}
