% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{main}
\alias{main}
\title{main.R}
\usage{
main(state, county)
}
\arguments{
\item{state}{A string representing the name of the state (e.g., "Wisconsin").s}

\item{county}{A string representing the name of the county (e.g., "Racine").}
}
\value{
None. Outputs are written to Excel files in the \code{DataFiles/OutputFiles} directory.
}
\description{
Main script to generate self-sufficiency data and related outputs
}
\details{
This script orchestrates the execution of various modules to generate self-sufficiency
data for a specified state and county. It performs the following tasks:
\itemize{
\item Retrieves job data and education requirements.
\item Processes food plans and housing costs.
\item Fetches and processes the Self Sufficiency Standard data.
\item Combines all data to calculate final self-sufficiency costs.
}
}
\note{
This script does not support Alaska or Hawaii.
}
