% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_self_suff_standard.R
\name{read_file}
\alias{read_file}
\title{Reads and processes the Self Sufficiency Standard file}
\usage{
read_file(file_path, county_self_sufficiency_standard)
}
\arguments{
\item{file_path}{A string representing the path to the Self Sufficiency Standard file.}

\item{county_self_sufficiency_standard}{A string representing the name of the county.}
}
\value{
A data frame containing processed data specific to the county.
}
\description{
This function reads the Self Sufficiency Standard Excel file, filters it for the specified
county, and processes the data by renaming columns, removing unnecessary columns, and
transforming specific columns to integers.
}
